Clean up and split files by purpose
